pyproject.toml
[project]
name = "sxml"
version = "0.1.11"
description = "Simple lxml pipeline"
keywords = ['lxml']
classifiers = [
   "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "PyYAML",
    "lxml",
    "cssselect",
]

[project.optional-dependencies]
extras = [
    "jsonpath-ng",
    "dateparser",
    "beautifulsoup4",
    "extruct",
]
test = [
    "freezegun",
    "lxml-stubs",
    "pytest",
    "pytest-cov",
    "pytest-mypy",
    "pytest-pylint",
    "pytest-flake8",
    "pytest-timeout",
    "types-PyYAML",
    "types-dateparser",
    "types-freezegun",
]

[project.urls]
Homepage="https://github.com/vmstarchenko/sxml"


[build-system]
requires = ["setuptools>=42.0", "wheel"]

# [options]
# packages=sxml
# zip_safe=true
# include_package_data=True

# [project.entry_points]
# console_scripts =
#     sxml = sxml.__main__:main

[tool.setuptools.packages.find]
include = ["sxml*"]

[tool.setuptools.package-data]
sxml = ["py.typed"]

[tool.coverage.run]
branch = true
include = "sxml/*"
omit= [
    "sxml/__main__.py"
]
relative_files = true

[tool.coverage.report]
show_missing = true

[tool.pytest]
flake8-ignore = [
    "tests/*",
]
flake8-max-line-length = 99

[tool.pytest.ini_options]
# addopts = "-ra -q"
filterwarnings = [
    "ignore::DeprecationWarning:pytest_pylint*",
    "ignore::DeprecationWarning:pylint*",
    "ignore::DeprecationWarning:pyRdfa*",
    "ignore::DeprecationWarning:flake8*",
    "ignore:The localize method is no longer necessary, as this time zone supports the fold attribute",
]
addopts = [
    "--flake8",
    "--pylint",
    "--mypy",
    # "-n", "4",
    # "-vv",
    "--cov=sxml", "--cov-report=xml", "--no-cov-on-fail",
    # "-k", "test_split",
    # "-r", "A",
]
testpaths = [
    "tests",
    "sxml",
]
